feat : add category context
