+++ src/components/CategoryCard.jsx
@@ -1,15 +1,14 @@
 import React from "react";
 
-export const CategoryCard = () => {
+export const CategoryCard = ({ category }) => {
+  const { categoryName, imgSrc, alt } = category;
   return (
     <>
       <div className="category-item radius-5 p-2">
-        <h2 className="h2 container-flex-center category-text">Whey Protein</h2>
-        <img
-          className="img-responsive"
-          src="https://fitmart-screens.netlify.app/assets/products/whey-protein/ON-Whey-Protein-5-lb.webp"
-          alt="whey-protein"
-        />
+        <h2 className="h2 container-flex-center category-text">
+          {categoryName}
+        </h2>
+        <img className="img-responsive" src={imgSrc} alt={alt} />
       </div>
     </>
   );
