+++ src/components/Loader/Loader.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const Loader = () => {
+  return <span className="loader"></span>;
+};
