+++ src/components/index.js
@@ -3,3 +3,4 @@ export { Footer } from "./Footer";
 export { Hero } from "./Hero";
 export { Navbar } from "./Navbar";
 export { TrumpCard } from "./TrumpCard";
+export { Loader } from "./Loader/Loader";
