+++ src/context/index.js
@@ -1,0 +1,1 @@
+export { CategoryProvider, useCategory } from "./category-context";
