+++ src/index.js
@@ -1,6 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import App from "./App";
+import { CategoryProvider } from "./context";
 import { makeServer } from "./server";
 
 // Call make Server
@@ -8,7 +9,9 @@ makeServer();
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <CategoryProvider>
+      <App />
+    </CategoryProvider>
   </React.StrictMode>,
   document.getElementById("root")
 );
