+++ src/pages/Home/Home.jsx
@@ -1,9 +1,12 @@
 import React from "react";
 import "./Home.css";
-import { Hero, TrumpCard } from "../../components";
+import { CategoryCard, Hero, TrumpCard } from "../../components";
 import { trumpCardObj } from "../../utils";
+import { useCategory } from "../../context";
 
 export const Home = () => {
+  const { categories } = useCategory();
+
   return (
     <>
       <Hero />
@@ -18,57 +21,9 @@ export const Home = () => {
         Featured Categories
       </h2>
       <div className="category m-4">
-        <a
-          href="/pages/category/protein.html"
-          className="category-item radius-5 p-2"
-        >
-          <h2 className="h2 container-flex-center category-text">
-            Whey Protein
-          </h2>
-          <img
-            className="img-responsive"
-            src="https://fitmart-screens.netlify.app/assets/products/whey-protein/ON-Whey-Protein-5-lb.webp"
-            alt="whey-protein"
-          />
-        </a>
-
-        <a
-          href="/pages/category/creatine.html"
-          className="category-item radius-5 p-2"
-        >
-          <h2 className="h2 container-flex-center category-text">Creatine</h2>
-          <img
-            className="img-responsive"
-            src="https://fitmart-screens.netlify.app/assets/products/creatine/dymatize-creatine.webp"
-            alt="creatine"
-          />
-        </a>
-
-        <a
-          href="/pages/category/gainer.html"
-          className="category-item radius-5 p-2"
-        >
-          <h2 className="h2 container-flex-center category-text">Gainer</h2>
-          <img
-            className="img-responsive"
-            src="https://fitmart-screens.netlify.app/assets/products/gainer/mb-gainer.webp"
-            alt="gainer"
-          />
-        </a>
-
-        <a
-          href="/pages/category/multi-vit.html"
-          className="category-item radius-5 p-2"
-        >
-          <h2 className="h2 container-flex-center category-text">
-            Multivitamins
-          </h2>
-          <img
-            className="img-responsive"
-            src="https://fitmart-screens.netlify.app/assets/products/multi-vitamins/mb-multi-vit.webp"
-            alt="multivitamins"
-          />
-        </a>
+        {categories.map((category) => (
+          <CategoryCard key={category._id} category={category} />
+        ))}
       </div>
     </>
   );
