+++ src/utils/trump-card-obj.js
